+++ svg3d/__init__.py
@@ -1,12 +1,20 @@
 from . import shaders
 from .svg3d import Engine, Mesh, _pad_arrays
-from .view import View, get_lookat_matrix, get_projection_matrix
+from .view import (
+    View,
+    get_lookat_matrix,
+    get_orthographic_matrix,
+    get_projection_matrix,
+    get_scene_rotation_matrix,
+)
 
 __all__ = [
     "_pad_arrays",
     "Engine",
     "get_lookat_matrix",
+    "get_orthographic_matrix",
     "get_projection_matrix",
+    "get_scene_rotation_matrix",
     "View",
     "shaders",
 ]